+++ lang/da/layouts.php
@@ -10,7 +10,7 @@ return [
     'email.subject' => 'Feedback :app_www',
     'modal.description' => 'Din feedback hjælper os med at forbedre. Del venligst dine tanker.',
     'modal.link_text' => 'Feedback',
-    'modal.success_message' => 'Tak for din feedback!',
+    'modal.success' => 'Tak, din besked er blevet sendt.',
     'modal.title' => 'Send feedback',
 
 ];