+++ lang/da/messages.php
@@ -1,6 +1,0 @@
-<?php
-
-return [
-    'success' => 'Tak for din feedback!',
-    'phone_validation' => 'Indtast venligst et gyldigt telefonnummer med 7-15 cifre.',
-];