+++ resources/views/components/button.blade.php
@@ -6,5 +6,5 @@
     {{ $attributes }}
     type="button"
     x-data
-    x-on:click.prevent="Livewire.dispatch('openFeedback', {id: '{{ $formId }}'})"
->{{ $slot ?? $formId }}</button>
+    x-on:click.prevent="Livewire.dispatch('openIgFeedback', {id: '{{ $formId }}'})"
+>{{ $slot->isNotEmpty() ? $slot : $formId }}</a>