+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Support custom success message.
+- Support custom validation error messages for fields.
+- Support dynamic attributes for fields.
+
 ## [4.1.0] - 2025-11-09
 
 _Stable release based on [4.1.0-rc.1]._
@@ -260,6 +268,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[Unreleased]: https://https://github.com/internetguru/laravel-feedback/compare/staging...dev
 [4.1.0]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.2...v4.1.0
 [4.1.0-rc.1]: https://github.com/internetguru/laravel-feedback/releases/tag/v4.0.2
 [4.0.2]: https://https://github.com/internetguru/laravel-feedback/compare/v4.0.1...v4.0.2